requirements-bot.txt
python-telegram-bot>=20.0
requests>=2.28
python-dotenv>=1.0